+++ examples/rainmaker/led_light/components/app_bridge/app_mesh_lite_comm.c
@@ -130,7 +130,7 @@ static const esp_mesh_lite_msg_action_t report_child_info_action[] = {
     {NULL, NULL, NULL} /* Must be NULL terminated */
 };
 
-static void root_timer_cb(void* param)
+static void root_timer_cb(TimerHandle_t timer)
 {
     if (esp_mesh_lite_get_level() > ROOT) {
         return;
@@ -167,7 +167,7 @@ static void root_timer_cb(void* param)
     xSemaphoreGive(report_child_info_mutex);
 }
 
-static void report_timer_cb(void* param)
+static void report_timer_cb(TimerHandle_t timer)
 {
     if (esp_mesh_lite_get_level() > ROOT) {
         app_rmaker_mesh_lite_report_child_info();